+++ lib/main.dart
@@ -1,5 +1,9 @@
 import 'dart:developer';
 
+import 'package:aa_travel_planner/FavoriteCard/favorite_card.dart';
+import 'package:aa_travel_planner/FavoritesDetailsScreen/favorites_details_screen.dart';
+import 'package:aa_travel_planner/FavoritesRepository/favorites_repository.dart';
+import 'package:aa_travel_planner/FavoritesScreen/favorites_screen.dart';
 import 'package:flutter/material.dart';
 
 void main() {
@@ -237,55 +241,6 @@ class TripRepository {
 
   List<Trip> getAllTrips() {
     return _trips;
-  }
-}
-
-class FavoritesRepository {
-  final List<Destination> _favoriteDestinations = [
-    Destination(
-      id: '4',
-      name: 'Rome',
-      country: 'Italy',
-      description:
-          'A city steeped in history, with ancient ruins and Renaissance art.',
-      imageUrl: 'rome.jpeg',
-    ),
-    Destination(
-      id: '5',
-      name: 'Sydney',
-      country: 'Australia',
-      description:
-          'Known for its Sydney Opera House, Harbour Bridge, and beautiful beaches.',
-      imageUrl: 'sydney.jpeg',
-    ),
-    Destination(
-      id: '6',
-      name: 'Rio de Janeiro',
-      country: 'Brazil',
-      description:
-          'Famous for its Carnival, Christ the Redeemer, and breathtaking beaches.',
-      imageUrl: 'rio_de_janeiro.jpeg',
-    ),
-  ];
-
-  List<Destination> getFavorites() {
-    return _favoriteDestinations;
-  }
-
-  void addFavorite(Destination destination) {
-    if (!_favoriteDestinations.contains(destination)) {
-      _favoriteDestinations.add(destination);
-      log('${destination.name} added to favorites');
-    }
-  }
-
-  void removeFavorite(Destination destination) {
-    _favoriteDestinations.remove(destination);
-    log('${destination.name} removed from favorites');
-  }
-
-  bool isFavorite(Destination destination) {
-    return _favoriteDestinations.contains(destination);
   }
 }
 
@@ -707,36 +662,6 @@ class ExploreDestinationsScreen extends StatelessWidget {
   }
 }
 
-class FavoritesScreen extends StatefulWidget {
-  final FavoritesRepository favoritesRepository;
-
-  const FavoritesScreen({super.key, required this.favoritesRepository});
-
-  @override
-  State<FavoritesScreen> createState() => _FavoritesScreenState();
-}
-
-class _FavoritesScreenState extends State<FavoritesScreen> {
-  @override
-  Widget build(BuildContext context) {
-    List<Destination> favorites = widget.favoritesRepository.getFavorites();
-
-    return ListView.builder(
-      itemCount: favorites.length,
-      itemBuilder: (context, index) {
-        final destination = favorites[index];
-        return FavoriteCard(
-          destination: destination,
-          favoritesRepository: widget.favoritesRepository,
-          onRemove: () {
-            setState(() {});
-          },
-        );
-      },
-    );
-  }
-}
-
 class SettingsScreen extends StatefulWidget {
   const SettingsScreen({super.key});
 
@@ -895,99 +820,6 @@ class TripDetailsScreen extends StatelessWidget {
                     ElevatedButton.styleFrom(backgroundColor: Colors.teal[600]),
                 child: const Text(
                   'Go Back',
-                  style: TextStyle(color: Colors.white),
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class FavoritesDetailsScreen extends StatelessWidget {
-  final Destination destination;
-  final FavoritesRepository favoritesRepository;
-
-  const FavoritesDetailsScreen({
-    super.key,
-    required this.destination,
-    required this.favoritesRepository,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('${destination.name} - Favorite Destination'),
-        backgroundColor: Colors.amber[700],
-      ),
-      body: SingleChildScrollView(
-        child: Padding(
-          padding: const EdgeInsets.all(16.0),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: <Widget>[
-              ClipRRect(
-                borderRadius: BorderRadius.circular(20),
-                child: Image.asset("assets/images/${destination.imageUrl}",
-                    fit: BoxFit.cover),
-              ),
-              const SizedBox(height: 20),
-              Center(
-                child: Container(
-                  padding:
-                      const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(10),
-                    color: Colors.amber[200],
-                  ),
-                  child: const Text(
-                    'FAVORITE DESTINATION',
-                    style: TextStyle(
-                        fontSize: 20,
-                        fontWeight: FontWeight.bold,
-                        color: Colors.white),
-                  ),
-                ),
-              ),
-              const SizedBox(height: 16),
-              Text(
-                destination.name,
-                style: TextStyle(
-                    fontSize: 24,
-                    fontWeight: FontWeight.bold,
-                    color: Colors.amber[900]),
-              ),
-              const SizedBox(height: 8),
-              Text(
-                destination.country,
-                style: TextStyle(
-                    fontSize: 18,
-                    fontWeight: FontWeight.w500,
-                    color: Colors.grey[600]),
-              ),
-              Divider(height: 30, color: Colors.amber[700]),
-              const Text(
-                'Description:',
-                style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
-              ),
-              const SizedBox(height: 8),
-              Text(
-                destination.description,
-                style: TextStyle(
-                    fontSize: 16, height: 1.5, color: Colors.grey[700]),
-              ),
-              const SizedBox(height: 24),
-              ElevatedButton(
-                onPressed: () {
-                  favoritesRepository.removeFavorite(destination);
-                  Navigator.pop(context);
-                },
-                style: ElevatedButton.styleFrom(backgroundColor: Colors.red),
-                child: const Text(
-                  'Remove from Favorites',
                   style: TextStyle(color: Colors.white),
                 ),
               ),
@@ -1180,97 +1012,6 @@ class DestinationCard extends StatelessWidget {
             ],
           ),
         ),
-      ),
-    );
-  }
-}
-
-class FavoriteCard extends StatelessWidget {
-  final Destination destination;
-  final VoidCallback onRemove;
-  final FavoritesRepository favoritesRepository;
-
-  const FavoriteCard({
-    super.key,
-    required this.destination,
-    required this.onRemove,
-    required this.favoritesRepository,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(15)),
-      elevation: 8,
-      color: Colors.amber[100],
-      margin: const EdgeInsets.symmetric(vertical: 10, horizontal: 16),
-      child: Stack(
-        children: [
-          InkWell(
-            onTap: () {
-              Navigator.push(
-                context,
-                MaterialPageRoute(
-                  builder: (context) => FavoritesDetailsScreen(
-                    destination: destination,
-                    favoritesRepository: favoritesRepository,
-                  ),
-                ),
-              ).then((_) => onRemove()); // Liste updaten bei Zurücknavigation
-            },
-            child: Padding(
-              padding: const EdgeInsets.all(16.0),
-              child: Row(
-                children: [
-                  ClipRRect(
-                    borderRadius: BorderRadius.circular(10),
-                    child: Image.asset("assets/images/${destination.imageUrl}",
-                        width: 80, height: 80, fit: BoxFit.cover),
-                  ),
-                  const SizedBox(width: 16),
-                  Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Text(destination.name,
-                          style: const TextStyle(
-                              fontSize: 20, fontWeight: FontWeight.bold)),
-                      const SizedBox(height: 8),
-                      Text(destination.country,
-                          style:
-                              TextStyle(fontSize: 14, color: Colors.grey[700])),
-                    ],
-                  ),
-                ],
-              ),
-            ),
-          ),
-          Positioned(
-            top: 8,
-            right: 8,
-            child: GestureDetector(
-              onTap: () {
-                favoritesRepository.removeFavorite(destination);
-                onRemove(); // Liste updaten bei Zurücknavigation
-              },
-              child: Container(
-                decoration: const BoxDecoration(
-                  shape: BoxShape.circle,
-                  gradient: LinearGradient(
-                    colors: [Colors.orange, Colors.red],
-                    begin: Alignment.topLeft,
-                    end: Alignment.bottomRight,
-                  ),
-                ),
-                padding: const EdgeInsets.all(6),
-                child: const Icon(
-                  Icons.favorite,
-                  color: Colors.white,
-                  size: 24,
-                ),
-              ),
-            ),
-          ),
-        ],
       ),
     );
   }
